Functions and tests renamed
