+++ decompy/reinforcementlearning/reward/RewardFunction.py
@@ -2,7 +2,7 @@
 # TODO: decide which things need to be static.
 import ctypes
 import pathlib
-#from decompy.RL.Model import Model
+from decompy.reinforcementlearning.model.Model import Model
 
 libreward_path = pathlib.PurePath.joinpath(pathlib.Path(__file__).resolve().parent, "libRewardFunction.so")
 print(libreward_path)
